Get module compatible with OCA Guidelines
