+++ l10n_ch_scan_bvr/__init__.py
@@ -1,25 +1,8 @@
-# -*- encoding: utf-8 -*-
-#
-#  l10n_ch_scan_bvr
-#
-#  Created by Nicolas Bessi and Vincent Renaville
-#
-#  Copyright (c) 2012 CamptoCamp. All rights reserved.
-##############################################################################
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# -*- coding: utf-8 -*-
+# Author: Nicolas Bessi, Vincent Renaville
+# (c) 2012 Camptocamp SA
+# (c) 2015 Alex Comba - Agile Business Group
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
+from . import models
 from . import wizard
-from . import partner
