+++ l10n_ch_scan_bvr/__openerp__.py
@@ -1,46 +1,12 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Nicolas Bessi, Vincent Renaville
-#    Copyright 2012 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Author: Nicolas Bessi, Vincent Renaville
+# (c) 2012 Camptocamp SA
+# (c) 2015 Alex Comba - Agile Business Group
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 {"name": "Switzerland - Scan ESR/BVR to create invoices",
- "description": """
-Scan ESR/BVR to create supplier invoices
-========================================
-
-This module works with C-channel or other OCR scanner.
-
-It helps you to create an invoice directly from the ESR/BVR Code.
-Find the menu entry called "Scan BVR" under Accounting -> Supplier.
-It open a popup from which you can scan the ESR/BVR number.
-It'll recognize the needed information and create an
-invoice for the right supplier.
-
-If you have completed the field "Default product supplier invoice"
-on the concerned supplier,
-it'll create a line with the proper amount and the given product.
-
-It currently supports BVR and BVR+
-
-""",
- "version": "1.0",
- "author": "Camptocamp,Odoo Community Association (OCA)",
+ "version": "8.0.1.0.0",
+ "author": "Camptocamp, Odoo Community Association (OCA)",
  "category": "Generic Modules/Others",
  "website": "http://www.camptocamp.com",
  "license": "AGPL-3",
@@ -48,7 +14,7 @@ It currently supports BVR and BVR+
              "l10n_ch_payment_slip",
              "account"],
  "data": ["wizard/scan_bvr_view.xml",
-          "partner_view.xml",
-          "bank_view.xml"],
+          "views/partner_view.xml",
+          "views/bank_view.xml"],
  'installable': True,
  }
