+++ l10n_ch_scan_bvr/models/__init__.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
-# Author: Nicolas Bessi
-# (c) 2012 Camptocamp SA
 # (c) 2015 Alex Comba - Agile Business Group
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from . import scan_bvr
+from . import partner
