+++ l10n_ch_scan_bvr/views/bank_view.xml
