+++ l10n_ch_scan_bvr/views/partner_view.xml
